some PR fixes
