+++ App.js
@@ -1,17 +1,6 @@
 import React, { Component } from 'react';
-import { createStackNavigator } from 'react-navigation';
 
-import TabNav from './src/app/screens/TabNav';
-import { ROUTES } from './src/constants/routes';
-
-const RootStack = createStackNavigator(
-  {
-    navbar: TabNav
-  },
-  {
-    initialRouteName: ROUTES.navbar
-  }
-);
+import RootStack from './src/app/screens';
 
 class App extends Component {
   render() {
