+++ src/app/screens/Book/index.js
@@ -1,13 +1,7 @@
 import React, { Component } from 'react';
 import { View, Text } from 'react-native';
 
-import { ROUTES } from '../../../constants/routes';
-
 class Book extends Component {
-  static navigationOptions = {
-    title: ROUTES.bookTitle
-  };
-
   render() {
     return (
       <View>
