+++ src/app/screens/Home/styles.js
