+++ src/constants/routes.js
@@ -1,8 +1,5 @@
 export const ROUTES = {
   todo: 'todo',
-  todoTitle: 'Todo',
   book: 'book',
-  bookTitle: 'Books',
-  navbar: 'navbar',
-  applicationTitle: 'Training APP'
+  home: 'home'
 };
