+++ src/constants/titles.js
@@ -1,0 +1,5 @@
+export const TITLES = {
+  todoTitle: 'Todo',
+  bookTitle: 'Books',
+  applicationTitle: 'Training APP'
+};
